[frontend] Adapt rounding and split useUserMetric to specific file because useAuth is need
